Changed fancySolveProblem() to create multiple ants, decay and deposit pheromone. Stopwatch now in minutes (input from the GUI).
Ants are given a random seed when they are created.
Ant class now has a DepositPheramone function, called for every ant after the round is complete.

I hope I didn't introduce this bug: The Ant public property Route (the indices of cities in the path that was found) seems to be set to null. But it may be that it never gets to the right block of code to set it equal to path, because the Console.WriteLine statement doesn't output what I have on line 39 in the else statement, so something else might be the issue.